requirements.txt
setuptools~=50.3.1
pymongo
pycrypto